requirements.txt
Django==4.2.8
django-bootstrap-v5==1.0.11
django-htmx==1.17.2
django-ninja==1.1.0
humanize==4.9.0
Jinja2==3.1.2
